% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function1.R
\name{DRG_box}
\alias{DRG_box}
\title{Boxplot for Payment Methods}
\usage{
DRG_box(df, payments)
}
\arguments{
\item{df}{a dataframe}

\item{payments}{three types payment methods:
average Medicare payments, the average total payment, or the average covered charges}
}
\value{
a list of boxplots of payment method by DRG code.
}
\description{
This function produces a boxplot based on DRG payment methods by DRG code:
average Medicare payments, the average total payment, or the average covered charges.
}
\examples{
DRG_box(df = DRG, payments = "Average.Medicare.Payments")
DRG_box(df = DRG, payments = "Average.Covered.Charges")
DRG_box(df = DRG, payments = "Average.Total.Payments")

}
