% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRG_statistics.R
\name{DRG_stat}
\alias{DRG_stat}
\title{Statistics of DRG}
\usage{
DRG_stat(method = c("mean", "median", "standard deviation"))
}
\arguments{
\item{method}{statistics}
}
\value{
to show \code{method} over all of the DRG codes for average Medicare payments
}
\description{
This function calculates statistics over all of the DRG codes for average Medicare payments
}
\examples{
DRG_stat("mean")

}
