% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function2.R
\name{medicare_stat}
\alias{medicare_stat}
\title{Statistical Calculation for Average Medicare Payments}
\usage{
medicare_stat(df, method)
}
\arguments{
\item{df}{a dataframe}

\item{method}{statistical methods: mean, median, and standard deviation}
}
\value{
A table of the statistical result over all of the DRG codes
for Average Medicare Payments based on the methods
}
\description{
This function calculates mean, median, and standard deviation
over all of the DRG codes for Average Medicare Payments.
}
\examples{
medicare_stat (DRG, mean)
medicare_stat (DRG, median)
medicare_stat (DRG, sd)

}
