% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRG_box.R
\name{DRG_box}
\alias{DRG_box}
\title{Boxplot of DRG}
\usage{
DRG_box(
  payment = c("Average.Covered.Charges", "Average.Total.Payments",
    "Average.Medicare.Payments")
)
}
\arguments{
\item{payment}{average money}
}
\value{
a boxplot to show \code{payment} of different DRG definition number
}
\description{
This function produces a boxplot to show \code{payment} of different DRG definition number
}
\examples{
DRG_box("Average.Covered.Charges")

}
